producers/noaa_phys/requirements.txt
requests
pandas
confluent-kafka
pyyaml
